+++ contracts/vault/src/msg.rs
@@ -1,31 +1,5 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, Binary, Uint128};
-
-/*** state ***/
-
-#[cw_serde]
-pub struct Config {
-    /// The denom we accept for staking (only native tokens)
-    pub denom: String,
-
-    /// The address of the local staking contract (where actual tokens go)
-    pub local_staking: Addr,
-}
-
-/// All values are in Config.denom
-#[cw_serde]
-pub struct Balance {
-    pub bonded: Uint128,
-    pub claims: Vec<LienAddr>,
-}
-
-#[cw_serde]
-pub struct LienAddr {
-    pub lienholder: Addr,
-    pub amount: Uint128,
-}
-
-/**** api ****/
+use cosmwasm_std::{Binary, Uint128};
 
 /// This is the info used to construct the native staking contract
 #[cw_serde]
